python-examples/ping-pong/pyproject.toml
[tool.poetry]
name = "chatgpt-clone-no-memory-langchain"
version = "0.1.0"
description = ""
authors = ["Kevin Piacentini <support@agentlabs.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
agentlabs-sdk = "^0.0.33"
pydantic = "^2.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
